Add API to inhibit idle actions
This introduces a platform-agnostic API for inhibiting idle actions such as automatic screen locking/turnoff.

We have code that covers this for multiple platforms in various apps, e.g. Itinerary and KDE Connect. Therefore we should move to a single codebase for it.

Currently it covers org.freedesktop.ScreenSaver and Android. KDE Connect contains code for Windows and macOS as well, this should eventually be added here too.