Deprecate single-arg signal KIdleTime::timeoutReached(int identifier)
Being an overload to the signal with the additional int msec argument,
connecting to it using a memberfunction pointer requires to specifcy
for the compiler which overload to use (e.g. with QOverload<>).
Instead of renaming one of the signals, the overload with less arguments
is dropped, as the multi-arg signal can be used for that need as well,
igoring the second argument for the slot.