Relicense file to LGPL-2.1-or-later
License change confirmed by manual check of authors and commit history
check via relicensecheck, with the following exception:

- lgplv2+: bluescreen_avenger@verizon.net jkt
bluescreen_avenger@verizon.net (   3 LOC): d613373

This change is a simple addition of an if-def constraint for X11, which
is below the copyright limit.

Moreover, Dario confirmed to license change in private mail, which is
stored in sysadmin archive.

NO_CHANGELOG