Add native wayland plugin
Currently this plugin exists in plasma code due to a dependency on
KWayland. This renders kidletime unusable outside of plasma even if the
implementation is used.
    
This patch provides support directly in kidletime using QtWayland
instead of relying on interim wrappers in KWayland that we hope to drop
in KF6.
    
Migration is handled implicitly. If the plasma plugin is available
either will be chosen at random. Then we will drop the plugin shipped by
plasma next release.