Rename AbstractSystemPoller -> KAbstractIdleTimePoller
Given this class is exposed both as exported library symbols as well
as via header includes (even if only "private"), adding some
namespacing prefix K as well as more context in the name specifier makes
the origin and purpose of the class more obvious in consumer context.

Also using _p postfix for the header for name pattern consistency with
other published private API in KF (and Qt).